Wait times now working in minutes
